creating 0.6.1 tag

+++ lib/SGL/Task/Install.php
@@ -91,13 +91,21 @@ class SGL_Task_CreateConfig extends SGL_Task
     function run($data)
     {
         $c = &SGL_Config::singleton($autoLoad = false);
+        $oldConf = $c->getAll(); // save old config on re-install
         $conf = $c->load(SGL_ETC_DIR . '/default.conf.dist.ini');
         $c->replace($conf);
+        $c->merge($oldConf); // overwrite with old values
 
         //  admin emails
         $c->set('email', array('admin' => $data['adminEmail']));
         $c->set('email', array('info' => $data['adminEmail']));
         $c->set('email', array('support' => $data['adminEmail']));
+
+        // correct db prefix
+        if (!empty($data['prefix']) && substr($data['prefix'], -1) != '_') {
+            // enforce underscore in prefix
+            $data['prefix'] .= '_';
+        }
 
         //  db details
         $c->set('db', array('prefix' => $data['prefix']));
@@ -164,6 +172,12 @@ class SGL_Task_CreateConfig extends SGL_Task
 
         //  and tz
         $_SESSION['install_timezone'] = $data['serverTimeOffset'];
+
+        //  store old prefix for tables drop
+        if (isset($oldConf['db']['prefix'])
+                && $oldConf['db']['prefix'] != $data['prefix']) {
+            $_SESSION['install_dbPrefix'] = $oldConf['db']['prefix'];
+        }
     }
 }
 
@@ -250,12 +264,15 @@ class SGL_Task_DefineTableAliases extends SGL_Task
     function run($data)
     {
         $c = &SGL_Config::singleton();
+
+        // get table prefix
+        $prefix = $c->get(array('db' => 'prefix'));
         foreach ($data['aModuleList'] as $module) {
             $tableAliasIniPath = SGL_MOD_DIR . '/' . $module  . '/data/tableAliases.ini';
             if (file_exists($tableAliasIniPath)) {
                 $aData = parse_ini_file($tableAliasIniPath);
                 foreach ($aData as $k => $v) {
-                    $c->set('table', array($k => $v));
+                    $c->set('table', array($k => $prefix . $v));
                 }
             }
         }
@@ -426,7 +443,14 @@ class SGL_Task_DropTables extends SGL_UpdateHtmlTask
             $statusText = 'dropping existing tables';
             $this->updateHtml('status', $statusText);
 
+            $c   = &SGL_Config::singleton();
             $dbh = & SGL_DB::singleton();
+
+            // set old db prefix if any
+            if (isset($_SESSION['install_dbPrefix'])) {
+                $currentPrefix = $c->get(array('db' => 'prefix'));
+                $c->set('db', array('prefix' => $_SESSION['install_dbPrefix']));
+            }
 
             //  drop 'sequence' table unless we're installing a module
             if ($this->conf['db']['type'] == 'mysql_SGL' && !array_key_exists('moduleInstall', $data)) {
@@ -470,13 +494,12 @@ class SGL_Task_DropTables extends SGL_UpdateHtmlTask
 
                     //  remove tablename in Config
                     if (isset($data['moduleInstall'])) {
-                        $c = &SGL_Config::singleton();
                         foreach ($aTableNames as $tableName) {
                             $c->remove(array('table', $tableName));
                         }
                         //  save
-                        $configFile = SGL_VAR_DIR . '/' . SGL_SERVER_NAME . '.conf.php';
-                        $ok = $c->save($configFile);
+                        $fileName = SGL_VAR_DIR . '/' . SGL_SERVER_NAME . '.conf.php';
+                        $ok = $c->save($fileName);
 
                         if (PEAR::isError($ok)) {
                             SGL_Install_Common::errorPush($ok);
@@ -486,6 +509,54 @@ class SGL_Task_DropTables extends SGL_UpdateHtmlTask
                 } else {
                     $this->updateHtml($module . '_drop', $this->noFile);
                 }
+            }
+            // remove translation tables and lang table
+            if (!array_key_exists('moduleInstall', $data)) {
+                $conf = $c->getAll();
+                if ($conf['translation']['container'] == 'db') {
+                    $statusText = 'dropping translation tables';
+                    $this->updateHtml('status', $statusText);
+                    $trans = &SGL_Translation::singleton('admin');
+                    $aLangs = $trans->getLangs('ids');
+                    if (!PEAR::isError($aLangs)) {
+                        // removeme
+                        if (empty($aLangs)) {
+                            // basically $aLangs should be a PEAR_Error instance
+                            // in that case, but calling method doesn't
+                            // return it
+                            SGL_Error::pop();
+                        }
+                        // dropping language tables
+                        foreach ($aLangs as $langId) {
+                            // force to drop translation table
+                            $ok = $trans->removeLang($langId, $force = true);
+                            if (PEAR::isError($ok, DB_ERROR_NOSUCHTABLE)) {
+                                SGL_Error::pop();
+                            }
+                        }
+                    } elseif (PEAR::isError($aLangs, DB_ERROR_NOSUCHTABLE)) {
+                        SGL_Error::pop();
+                    }
+                    // drop language table
+                    $langTable = &$trans->storage->options['langs_avail_table'];
+                    $query = 'DROP TABLE ' . $dbh->quoteIdentifier($langTable);
+                    $ok = $dbh->query($query);
+                    if (PEAR::isError($ok, DB_ERROR_NOSUCHTABLE)) {
+                        SGL_Error::pop();
+                    }
+
+                    // removeme: it looks like a hack
+                    if (isset($currentPrefix)) {
+                        $pattern   = "/^{$conf['db']['prefix']}/";
+                        $langTable = preg_replace($pattern, '', $langTable);
+                        $langTable = $currentPrefix . $langTable;
+                    }
+                }
+            }
+
+            // restore db prefix
+            if (isset($currentPrefix)) {
+                $c->set('db', array('prefix' => $currentPrefix));
             }
         }
     }
@@ -606,8 +677,8 @@ class SGL_Task_RemoveDefaultData extends SGL_Task
 {
     function run($data)
     {
-        require_once SGL_MOD_DIR . '/default/classes/DA_Default.php';
-        $da = & DA_Default::singleton();
+        require_once SGL_MOD_DIR . '/default/classes/DefaultDAO.php';
+        $da = & DefaultDAO::singleton();
 
         //  get perms associated with module
         $aPermNames = $da->getPermNamesByModuleId($data['moduleId']);
@@ -645,9 +716,9 @@ class SGL_Task_LoadBlockData extends SGL_UpdateHtmlTask
                 if (file_exists($modulePath . $this->filename4)) {
                     $result = SGL_Sql::parse($modulePath . $this->filename4, 0, array('SGL_Sql', 'execute'));
                     $displayHtml = $result ? $this->success : $this->failure;
-                    $this->updateHtml($module . '_dataBlock', $displayHtml);
+                    $this->updateHtml($module . '_dataSample', $displayHtml);
                 } else {
-                    $this->updateHtml($module . '_dataBlock', $this->noFile);
+                    $this->updateHtml($module . '_dataSample', $this->noFile);
                 }
             }
         }
@@ -810,7 +881,6 @@ class SGL_Task_LoadTranslations extends SGL_UpdateHtmlTask
 {
     function run($data)
     {
-        $configFile = SGL_VAR_DIR . '/' . SGL_SERVER_NAME . '.conf.php';
         $c = &SGL_Config::singleton();
         $aLangOptions = SGL_Util::getLangsDescriptionMap();
 
@@ -831,25 +901,27 @@ class SGL_Task_LoadTranslations extends SGL_UpdateHtmlTask
                 ? implode(',', str_replace('-', '_', $data['installLangs']))
                 : '';
             $c->set('translation', array('installedLanguages' => $langString));
-            $ok = $c->save($configFile);
-            if (PEAR::isError($ok)) {
-                SGL_Install_Common::errorPush($ok);
-            }
 
             //  iterate through languages adding to langs table
             foreach ($data['installLangs'] as $aLang) {
                 $globalLangFile = $availableLanguages[$aLang][1] .'.php';
                 $langID = str_replace('-', '_', $aLang);
-                $encoding = substr($aLang, strpos('-', $aLang));
-                $langData  = array(
-                    'lang_id' => $langID,
-                    'table_name' => $this->conf['table']['translation'] .'_'. $langID,
-                    'meta' => '',
-                    'name' => $aLangOptions[$aLang],
-                    'error_text' => 'not available',
-                    'encoding' => $encoding
-                     );
-                $result = $trans->addLang($langData);
+
+                // skip language creation during module install
+                if (empty($data['skipLangTablesCreation'])) {
+                    $prefix = $this->conf['db']['prefix'] .
+                        $this->conf['translation']['tablePrefix'] . '_';
+                    $encoding = substr($aLang, strpos('-', $aLang));
+                    $langData = array(
+                        'lang_id' => $langID,
+                        'table_name' => $prefix . $langID,
+                        'meta' => '',
+                        'name' => $aLangOptions[$aLang],
+                        'error_text' => 'not available',
+                        'encoding' => $encoding
+                    );
+                    $result = $trans->addLang($langData);
+                }
 
                 //  iterate through modules
                 foreach ($data['aModuleList'] as $module) {
@@ -890,11 +962,15 @@ class SGL_Task_LoadTranslations extends SGL_UpdateHtmlTask
         } else {
             //  set installed languages
             $installedLangs = (is_array($aLangOptions))
-                ? implode(',', str_replace('-', '_', array_keys($aLangOptions)))
+                ? str_replace('-', '_', implode(',', array_keys($aLangOptions)))
                 : '';
 
             $c->set('translation', array('installedLanguages' => $installedLangs));
-            $ok = $c->save($configFile);
+        }
+        $fileName = SGL_VAR_DIR . '/' . SGL_SERVER_NAME . '.conf.php';
+        $ok = $c->save($fileName);
+        if (PEAR::isError($ok)) {
+            SGL_Install_Common::errorPush($ok);
         }
     }
 }
@@ -980,18 +1056,22 @@ class SGL_Task_CreateFileSystem extends SGL_Task
 
         //  pass paths as arrays to avoid widows space parsing prob
         //  create cache dir
-        $cacheDir = System::mkDir(array(SGL_CACHE_DIR));
-        @chmod($cacheDir, 0777);
+        if (!is_dir(SGL_CACHE_DIR)) {
+            $cacheDir = System::mkDir(array(SGL_CACHE_DIR));
+            @chmod($cacheDir, 0777);
 
-        if (!($cacheDir)) {
-            SGL_Install_Common::errorPush(PEAR::raiseError('Problem creating cache dir'));
+            if (!($cacheDir)) {
+                SGL_Install_Common::errorPush(PEAR::raiseError('Problem creating cache dir'));
+            }
         }
 
         //  create entities dir
-        $entDir = System::mkDir(array(SGL_ENT_DIR));
-        @chmod($entDir, 0777);
-        if (!($entDir)) {
-            SGL_Install_Common::errorPush(PEAR::raiseError('Problem creating entity dir'));
+        if (!is_dir(SGL_ENT_DIR)) {
+            $entDir = System::mkDir(array(SGL_ENT_DIR));
+            @chmod($entDir, 0777);
+            if (!($entDir)) {
+                SGL_Install_Common::errorPush(PEAR::raiseError('Problem creating entity dir'));
+            }
         }
 
         //  create tmp dir, mostly for sessions
@@ -1038,6 +1118,22 @@ class SGL_Task_CreateDataObjectEntities extends SGL_Task
             $ok = unlink($keysFile);
         }
 
+        // drop old entities on re-install
+        if (isset($_SESSION['install_dbPrefix'])) {
+            if (is_writable(SGL_ENT_DIR)) {
+                if ($dh = opendir(SGL_ENT_DIR)) {
+                    $prefix = $_SESSION['install_dbPrefix'];
+                    while (($file = readdir($dh)) !== false) {
+                        if ($file != '.' && $file != '..'
+                                && substr($file, -3) == 'php'
+                                && substr($file, 0, strlen($prefix)) == ucfirst($prefix)) {
+                            $ok = unlink(SGL_ENT_DIR . '/' . $file);
+                        }
+                    }
+                }
+            }
+        }
+
         $generator = new DB_DataObject_Generator();
         $generator->start();
         $out = ob_get_contents();
@@ -1060,16 +1156,18 @@ class SGL_Task_CreateDataObjectLinkFile extends SGL_Task
         $c = &SGL_Config::singleton();
         $conf = $c->getAll();
 
-        // remove original dbdo links file
+        // original dbdo links file
         $linksFile = SGL_ENT_DIR . '/' . $conf['db']['name'] . '.links.ini';
 
-        //  only remove when not installing modules, ie for sgl-rebuild
-        if (empty($data['moduleInstall'])) {
-            if (is_file($linksFile) && is_writable($linksFile)) {
-                $aOrigData = parse_ini_file($linksFile, true);
+        // read existing data if any
+        if (is_readable($linksFile)) {
+            $aOrigData = parse_ini_file($linksFile, true);
+            // only remove when not installing modules, ie for sgl-rebuild
+            if (empty($data['moduleInstall']) && is_writable($linksFile)) {
                 unlink($linksFile);
             }
         }
+
         $linkData = '';
         foreach ($data['aModuleList'] as $module) {
             $linksPath = SGL_MOD_DIR . '/' . $module  . '/data/dataobjectLinks.ini';
@@ -1086,14 +1184,24 @@ class SGL_Task_CreateDataObjectLinkFile extends SGL_Task
                 //  compare with existing data if there is any
                 if (!empty($aOrigData)) {
                     foreach ($aNewData as $key => $aValues) {
-                        if (array_key_exists($key, $aOrigData)) {
+                        $tableName = $conf['db']['prefix'] . $key;
+                        if (array_key_exists($tableName, $aOrigData)) {
                             //  key already exists, so return instead of adding it
                             return;
                         }
                     }
                 }
             }
-            if (is_writable($linksFile)) {
+            // we don't forget about prefixes
+            if (!empty($conf['db']['prefix'])) {
+                // prefix containers
+                $linkData = preg_replace('/\[(\w+)\]/i',
+                    '[' . SGL_Sql::addTablePrefix('$1') . ']',  $linkData);
+                // prefix references
+                $linkData = preg_replace('/(\w+):/i',
+                    SGL_Sql::addTablePrefix('$1') . ':' , $linkData);
+            }
+            if (is_writable($linksFile) || !file_exists($linksFile)) {
                 if (!$handle = fopen($linksFile, 'a+')) {
                     SGL_Install_Common::errorPush(
                         PEAR::raiseError('could not open links file for writing'));
@@ -1116,9 +1224,28 @@ class SGL_Task_SymLinkWwwData extends SGL_Task
             if (file_exists($wwwDir)) {
                 if (is_writable(SGL_WEB_ROOT)) {
                     if (file_exists(SGL_WEB_ROOT . "/$module")) {
-                        unlink(SGL_WEB_ROOT . "/$module");
+                        PEAR::raiseError('A www directory was detected in ' .
+                            ' one of the modules therefore an attempt to create ' .
+                            ' a corresponding symlink was made ' .
+                            ' but the symlink already exists ' .
+                            ' in seagull/www');
+                    } else {
+
+                        // windows
+                        if (strpos(PHP_OS, 'WIN') !== false) {
+
+                            // if linkd binary is present
+                            $ok = symlink($wwwDir, SGL_WEB_ROOT . "/$module");
+
+                            //  otherwise just copy
+                            if (!$ok) {
+                                require_once SGL_CORE_DIR . '/File.php';
+                                $success = SGL_File::copyDir($wwwDir, SGL_WEB_ROOT . "/$module");
+                            }
+                        } else {
+                            $ok = symlink($wwwDir, SGL_WEB_ROOT . "/$module");
+                        }
                     }
-                    $ok = symlink($wwwDir, SGL_WEB_ROOT . "/$module");
                 } else {
                     PEAR::raiseError('A www directory was detected in one of the modules '.
                     ' but the required webserver' .
@@ -1136,11 +1263,23 @@ class SGL_Task_UnLinkWwwData extends SGL_Task
     {
         foreach ($data['aModuleList'] as $module) {
             $wwwDir = SGL_MOD_DIR . '/' . $module  . '/www';
-            if (file_exists($wwwDir)) {
+            // if we're windows
+            if ((strpos(PHP_OS, 'WIN') !== false) && is_dir(SGL_WEB_ROOT . "/$module")) {
+                require_once SGL_CORE_DIR . '/File.php';
+                if (readlink(SGL_WEB_ROOT . "/$module")) {
+                    SGL_File::rmDir(SGL_WEB_ROOT . "/$module");
+                } else {
+                    SGL_File::rmDir(SGL_WEB_ROOT . "/$module", '-r');
+                }
+            } elseif (file_exists($wwwDir)) {
                 if (is_writable(SGL_WEB_ROOT)) {
                     if (file_exists(SGL_WEB_ROOT . "/$module")) {
                         unlink(SGL_WEB_ROOT . "/$module");
                     }
+                } else {
+                    PEAR::raiseError('An attempt to remove an existing ' .
+                        ' symlink failed, the webserver no longer has ' .
+                        ' required write perms on seagull/www dir');
                 }
             }
         }
@@ -1456,4 +1595,20 @@ PHP;
         }
     }
 }
+
+if (strpos(PHP_OS, 'WIN') !== false) {
+    if (!function_exists('symlink')) {
+        function symlink($target, $link) {
+            exec("linkd " . $link . " " . $target, $ret);
+            return $ret;
+        }
+    }
+    if (!function_exists('readlink')) {
+        function readlink($link) {
+            exec('linkd ' . $link, $ret);
+            return $ret;
+        }
+    }
+}
+
 ?>
